temp: 커밋 수정
잘못올린 커밋 수정

+++ External/Include/Script/CCoolDownUIScript.h
@@ -5,7 +5,7 @@ class CCoolDownUIScript :
     public CScript
 {
 private:
-    double m_dCoolDown;
+    float m_dCoolDown;
 
 
 public:
